+++ lib/src/async/async_data.dart
@@ -1,13 +1,9 @@
 import 'dart:async';
 
 import 'package:alien_signals/alien_signals.dart' as alien;
+import 'package:alien_signals/system.dart' as alien;
 import 'package:flutter/widgets.dart';
-
-import '../core/batch.dart';
-import '../core/effect.dart';
-import '../core/signal.dart';
-import '../core/untrack.dart';
-import '../core/watch.dart';
+import 'package:oref/oref.dart' as oref;
 
 enum AsyncStatus { idle, pending, success, error }
 
@@ -16,13 +12,13 @@ abstract interface class AsyncData<T> {
   abstract AsyncError? error;
   abstract T? data;
 
-  void stop();
+  void dispose();
   Future<T?> refresh();
 
   R when<R>({
     BuildContext? context,
     required R Function(T?) idle,
-    required R Function() pending,
+    required R Function(T?) pending,
     required R Function(T? data) success,
     required R Function(AsyncError? error) error,
   });
@@ -34,9 +30,12 @@ AsyncData<T> useAsyncData<T>(
   ValueGetter<T?>? defaults,
 }) {
   final executor = _AsyncDataExecutor<T>(
-    data: signal<T?>(context, defaults != null ? untrack(defaults) : null),
-    error: signal<AsyncError?>(context, null),
-    status: signal(context, AsyncStatus.idle),
+    data: oref.signal<T?>(
+      context,
+      defaults != null ? oref.untrack(defaults) : null,
+    ),
+    error: oref.signal<AsyncError?>(context, null),
+    status: oref.signal(context, AsyncStatus.idle),
     handler: handler,
   );
 
@@ -45,11 +44,7 @@ AsyncData<T> useAsyncData<T>(
   return _AsyncDataImpl(executor, context);
 }
 
-typedef _Signal<T> = T Function([T?, bool]);
-
 class _AsyncDataExecutor<T> {
-  static final store = Expando<alien.ReactiveNode>('oref:async_data');
-
   _AsyncDataExecutor({
     required this.data,
     required this.error,
@@ -57,36 +52,33 @@ class _AsyncDataExecutor<T> {
     required this.handler,
   });
 
-  final _Signal<T?> data;
-  final _Signal<AsyncError?> error;
-  final _Signal<AsyncStatus> status;
+  final oref.WritableSignal<T?> data;
+  final oref.WritableSignal<AsyncError?> error;
+  final oref.WritableSignal<AsyncStatus> status;
   final ValueGetter<FutureOr<T>> handler;
 
-  late final VoidCallback stop;
   late final alien.ReactiveNode node;
 
   bool isInitialized = false;
   late Completer<T?> completer = Completer.sync()..complete(null);
 
   void ensureInitialized(BuildContext? context) {
-    stop = effect(context, () {
+    final prevIsInitialized = isInitialized;
+    final effect = oref.effect(context, () {
       if (!isInitialized) {
         isInitialized = true;
-        if (context != null) {
-          store[context] = alien.getCurrentSub()!;
-        }
       }
 
       Future.microtask(schedule);
     });
 
-    if (context != null) {
-      node = store[context]!;
+    if (!prevIsInitialized) {
+      node = effect as alien.ReactiveNode;
     }
   }
 
   void schedule() async {
-    if (untrack(status) == AsyncStatus.pending) {
+    if (oref.untrack(status.call) == AsyncStatus.pending) {
       return;
     } else if (completer.isCompleted) {
       completer = Completer();
@@ -96,30 +88,28 @@ class _AsyncDataExecutor<T> {
 
     try {
       final result = await scoped(handler);
-      batch(() {
-        this
-          ..data(result)
-          ..status(AsyncStatus.success);
+      oref.batch(() {
+        data(result);
+        status(AsyncStatus.success);
       });
 
       completer.complete(result);
     } catch (error, stackTrace) {
-      batch(() {
-        this
-          ..error(AsyncError(error, stackTrace))
-          ..status(AsyncStatus.error);
+      oref.batch(() {
+        this.error(AsyncError(error, stackTrace));
+        status(AsyncStatus.error);
       });
 
       completer.completeError(error, stackTrace);
     }
   }
 
-  R scoped<R>(R Function() run) {
-    final prevSub = alien.setCurrentSub(node);
+  Future<R> scoped<R>(FutureOr<R> Function() run) async {
+    final prevSub = alien.setActiveSub(node);
     try {
-      return run();
+      return await run();
     } finally {
-      alien.setCurrentSub(prevSub);
+      alien.setActiveSub(prevSub);
     }
   }
 }
@@ -146,7 +136,7 @@ class _AsyncDataImpl<T> implements AsyncData<T> {
   AsyncStatus get status => executor.status();
 
   @override
-  set status(AsyncStatus value) => executor.status(value);
+  set status(AsyncStatus value) => executor.status(value, true);
 
   @override
   Future<T?> refresh() async {
@@ -158,25 +148,27 @@ class _AsyncDataImpl<T> implements AsyncData<T> {
   }
 
   @override
-  void stop() => executor.stop();
+  void dispose() {
+    (executor.node as oref.Effect).dispose();
+  }
 
   @override
   R when<R>({
     BuildContext? context,
     required R Function(T?) idle,
-    required R Function() pending,
+    required R Function(T?) pending,
     required R Function(T? data) success,
     required R Function(AsyncError? error) error,
   }) {
     context ??= this.context;
     R builder() => switch (status) {
       AsyncStatus.idle => idle(data),
-      AsyncStatus.pending => pending(),
+      AsyncStatus.pending => pending(data),
       AsyncStatus.success => success(data),
       AsyncStatus.error => error(this.error),
     };
     if (context != null) {
-      return watch(context, builder);
+      return oref.watch(context, builder);
     }
 
     return builder();